requirements.txt
numpy
pandas
scikit-learn
scipy
joblib
sentence-transformers
python-dotenv
torch
transformers
streamlit
pytest
httpx
